requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
streamlit==1.38.0
pandas==2.2.2
numpy==2.1.1
requests==2.32.3
python-dateutil==2.9.0.post0
pydantic==2.9.2
PyYAML==6.0.2
sqlalchemy==2.0.35
pytest==8.3.3
